Add the leopard complex (appaloosa): LP + PATN1 + PATN2
Three new loci and the model's first cross-locus-reading gene. LP (TRPM1)
alone gives the appaloosa characteristics - roaning that spares the bony
parts, a scatter of white spots, striped hooves and a white-rimmed eye.
The two modifier loci PATN1 and PATN2 mean nothing on a horse with no LP;
LeopardGene.expressionIn reads them (plus LP zygosity) to pick one of eight
outcomes: mottled, varnish-roan, leopard, fewspot, spotted blanket, snowcap,
semi-leopard.

New infrastructure:
- Gene.coatDependsOn() -> List<String>. The PATN modifiers paint nothing, so
  affectsCoat() is false and they'd be dropped from Genotype.coatCode(), which
  would collide a LP/LP PATN1/PATN1 horse with a plain LP/LP one in the texture
  cache. LeopardGene names both; coatCode() folds their alleles in, but only
  for a horse whose LP actually paints, so a PATN carrier with no LP still
  shares its texture (golden test asserts byte-identity). BreedFounder treats
  a coatDependsOn gene like a coat gene (forced wild unless the breed names it).
- GenotypeCatalog.size()/get() widened int -> long. The two-outcome coat gene
  doubled the reduced catalogue to 1,849,688,066, past Integer.MAX_VALUE. No
  production code indexes the catalogue (random pens), so only tests needed
  touching. entries() stays a List and now truncates at MAX_VALUE.

Painters (phase 1, off the expressing LP copy's seed): paintSpotted (Worley
spot field), paintBlanket (rump-forward white sheet + interior spots),
paintVarnish (asymmetric hair mix sparing the mane/tail/lower legs). Striped
hooves + white sclera in the overlay phase via CoatOverlayContribution.
LP/LP carries CSNB as an informational Condition (HealthContribution).

Founders: LP ~1/40 per allele; PATN drawn onto a founder only if it already
rolled LP (AppaloosaModifierGene reads the FounderContext). Appaloosa,
Knabstrupper, Kiger Mustang, Mustang and Shetland roll real LP/PATN pools;
the light-roan stand-in and the "NOT BUILT" notes are gone.

Also: the custom horse spawn egg's Breed button is now a scrollable dropdown
(was a cycle button), a parallel of the existing many-allele allele dropdown.

Tests: LeopardComplexGeneTest (14). coat-golden.txt regenerated - inserting
LEOPARD at priority 73 shifts Epigenome.random's positional per-gene seed
stream for every gene after it, so all non-deterministic kit/mitf/pax3/ednrb
rows moved; deterministic coats are byte-identical. WhitePatternGenesTest
.averageWhite widened 3 -> 8 seeds (sabinoWhiteIsAtLeastNinetyPerCentWhite was
a 3-seed knife-edge, not a KIT regression). SpecGeneTest.BUILT_IN_GENES 43->46,
HorseTraitsTest contributor count 18->19 (CSNB). CoatSampleTool: 8 lp_* samples,
and the long-broken chestnut_test sample (deleted `test` gene) removed.

405 common tests green; neoforge builds; runServer boots clean (48 segments).
Old saves will not parse.

+++ common/src/test/java/com/example/horsegenetics/common/genetics/GenotypeCatalogTest.java
@@ -143,7 +143,7 @@ class GenotypeCatalogTest {
      */
     @Test
     void eachMaskingCombinationContributesExactlyOneWildTypeEntry() {
-        int size = GenotypeCatalog.size();
+        long size = GenotypeCatalog.size();
         int maskingCombinations = 0;
         for (Gene gene : Genes.codeOrder()) {
             for (AllelePair pair : GenotypeCatalog.distinctPairsOf(gene)) {
@@ -154,7 +154,7 @@ class GenotypeCatalogTest {
         }
         assertTrue(maskingCombinations > 0, "the model should still have masking combinations to check");
 
-        for (int i = size - maskingCombinations; i < size; i++) {
+        for (long i = size - maskingCombinations; i < size; i++) {
             Genotype only = GenotypeCatalog.get(i);
             Gene masker = null;
             for (Gene gene : Genes.codeOrder()) {
@@ -176,7 +176,7 @@ class GenotypeCatalogTest {
         Random rng = new Random(20260903L);
         long plain = size - maskingCombinations;
         for (int n = 0; n < 4000; n++) {
-            Genotype g = GenotypeCatalog.get((int) (rng.nextDouble() * plain));
+            Genotype g = GenotypeCatalog.get((long) (rng.nextDouble() * plain));
             for (Gene gene : Genes.codeOrder()) {
                 assertFalse(gene.expressionOf(g.pair(gene)).masks(),
                         "an unmasked entry carries a masking combination: " + g.toCode());
@@ -194,7 +194,7 @@ class GenotypeCatalogTest {
     @Test
     void everyEntryIsDistinctAndRoundTripsThroughTheCode() {
         Set<String> codes = new HashSet<>();
-        for (int i : sampleIndices(6000)) {
+        for (long i : sampleIndices(6000)) {
             Genotype g = GenotypeCatalog.get(i);
             String code = g.toCode();
             assertTrue(codes.add(code), "duplicate genotype at index " + i + ": " + code);
@@ -206,7 +206,7 @@ class GenotypeCatalogTest {
     @Test
     void noTwoEntriesShareADisplayLabel() {
         Set<String> labels = new HashSet<>();
-        for (int i : sampleIndices(6000)) {
+        for (long i : sampleIndices(6000)) {
             String label = GeneCodeDisplay.shortForm(GenotypeCatalog.get(i));
             assertTrue(labels.add(label), "duplicate label: " + label);
         }
@@ -229,7 +229,7 @@ class GenotypeCatalogTest {
     @Test
     void everyLabelWrapsOntoThreeSignLinesWithoutLosingAnything() {
         int widestLast = 0;
-        for (int i : sampleIndices(6000)) {
+        for (long i : sampleIndices(6000)) {
             Genotype g = GenotypeCatalog.get(i);
             List<String> lines = GeneCodeDisplay.wrap(g, 3, 15);
             assertTrue(lines.size() <= 3, "too many lines for " + GeneCodeDisplay.shortForm(g) + ": " + lines);
@@ -247,21 +247,21 @@ class GenotypeCatalogTest {
      * odometer's low digits and the masked tail live - plus a seeded random
      * spread over everything in between.
      */
-    private static int[] sampleIndices(int n) {
-        int size = GenotypeCatalog.size();
-        Set<Integer> picked = new LinkedHashSet<>();
-        int ends = Math.min(size / 2, n / 6);
-        for (int i = 0; i < ends; i++) {
+    private static long[] sampleIndices(int n) {
+        long size = GenotypeCatalog.size();
+        Set<Long> picked = new LinkedHashSet<>();
+        int ends = (int) Math.min(size / 2, n / 6);
+        for (long i = 0; i < ends; i++) {
             picked.add(i);
             picked.add(size - 1 - i);
         }
         Random rng = new Random(4242L);
         while (picked.size() < Math.min(n, size)) {
-            picked.add((int) (rng.nextDouble() * size));
+            picked.add((long) (rng.nextDouble() * size));
         }
-        int[] out = new int[picked.size()];
+        long[] out = new long[picked.size()];
         int k = 0;
-        for (int i : picked) {
+        for (long i : picked) {
             out[k++] = i;
         }
         return out;
